feat(modal): add close button
