+++ src/css/styles.css
@@ -266,6 +266,7 @@ footer p {
   background-position: center;
   background-size: cover;
   border-radius: 5px;
+  position: relative;
 }
 
 .meal__title {
@@ -288,6 +289,24 @@ footer p {
 
 .ingredients__list {
   column-count: 3;
+}
+
+.close__button {
+  position: absolute;
+  height: 25px;
+  width: 25px;
+  background-color: var(--white-color);
+  border: none;
+  color: inherit;
+  border: none;
+  padding: 0;
+  font: inherit;
+  cursor: pointer;
+  outline: inherit;
+  top: 15px;
+  right: 15px;
+  border-radius: 50%;
+  text-align: center;
 }
