+++ src/index.html
@@ -6,6 +6,8 @@
    <meta http-equiv="X-UA-Compatible" content="IE=edge">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <title>Recipe finder</title>
+   <link rel="stylesheet"
+      href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@48,400,0,0" />
 </head>
 
 <body>
@@ -126,8 +128,12 @@
    </article>
    <div class="modal">
       <div class="meal">
-         
          <div class="meal__image">
+            <button class="close__button">
+               <span class="material-symbols-outlined">
+                  close
+               </span>
+            </button>
             <div class="meal__closing-button"></div>
          </div>
          <div class="meal__details">
